+++ src/server/game/Entities/Item/Item.h
@@ -35,6 +35,12 @@ struct ItemSetEffect
     SpellInfo const* spells[8];
 };
 
+enum InventorySlot
+{
+    NULL_BAG    = 0,
+    NULL_SLOT   = 255
+};
+
 // EnumUtils: DESCRIBE THIS
 enum InventoryResult : uint8
 {
